refactor(network): split out into utils-network package (#3537)
