+++ packages/@ionic/utils-network/.npmrc
@@ -1,0 +1,1 @@
+package-lock=false
