+++ packages/@ionic/utils-network/README.md
@@ -1,0 +1,1 @@
+# @ionic/utils-network
