+++ packages/@ionic/utils-network/tslint.js
@@ -1,0 +1,3 @@
+module.exports = {
+  "extends": "../../../tslint.base",
+};
